serverside/requirements.txt
streamlit
google-generativeai
python-dotenv==1.0.0
langchain
langchain-community
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pandas
openpyxl
python-Levenshtein
python-multipart
pydantic
fastapi 
uvicorn
websockets
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
werkzeug==3.0.1
eventlet==0.35.2
